Actor Structs
This one was painful... I had to accept I wouldn't be able to do as much
crazy stuff with Rust syntax, systematically running into the "uNsTaBLe
fEaTuRe" wall (yes, I use dead memes).
The foundations are there, there's certainly room for improvement, but
I've got the Actor Model I wanted for Rust. The pace of publications
will slow down, now that Absinthe has the features it needs for the
project it was created for.
But don't worry, this project will live on, for a long time.